# docs/api/openapi.yaml
openapi: 3.2.0
...
info:
  title: Imperio Industrial — API pública
  version: 1.5.0
  summary: API REST del MMO de simulación económica, industrial y logística "Imperio Industrial".
  description: |
    API pública del servidor autoritativo de **Imperio Industrial** (GDD/SAD v1.2).

    ## Principios del contrato de API

    - **Una sola API para humanos y bots** (igualdad de API literal, ADR-010): mismos
      endpoints y mismos rate limits lógicos. El motor no distingue el origen de un comando.
    - **Servidor autoritativo**: el cliente solo envía intenciones; toda validación
      (espacio, fondos, stock, plazos) ocurre server-side.
    - **REST para operaciones no urgentes**; los eventos en tiempo real del área de interés
      (movimiento de vehículos, alertas configuradas) se distribuyen por **WebSocket**
      a través del Notification/Event Gateway y quedan fuera de este documento.
    - **El tablón global es pull, no push**: se consulta bajo demanda con filtros
      (producto, ubicación, precio, plazo); nunca hay push mundial de mercado.
    - **Sim-time en el contrato**: todo plazo de dominio (contratos, producción, viajes)
      se define y transmite en sim-time (segundos desde el génesis del mundo, ratio 24×).
      La traducción a tiempo real es responsabilidad exclusiva de la UI. Las dos únicas
      mecánicas en tiempo real son la ventana de sorteo y el cooldown anti-parpadeo.
...
      (invariante del ledger).
...
      distinguibles.
...

    ## Estructura de respuestas

...

...
servers:
...
security:
  - sessionToken: []
...
paths:
...
  /world/buildings:
    get:
      operationId: listBuildings
      tags: [world, production]
      summary: Edificios propios
      parameters:
        - name: region_id
          in: query
          schema:
            $ref: '#/components/schemas/RegionId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BuildingStatus'
        - name: building_type_id
          in: query
          schema:
            $ref: '#/components/schemas/BuildingTypeId'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Edificios de la corporación.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Building'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
    post:
      operationId: createBuilding
      tags: [world, production]
      summary: Construir un edificio
      description: |
        Inicia la construcción sobre una concesión propia. Validación server-side de
        emplazamiento (espacio, acceso, recursos) contra las `placement_rules` del tipo
        y el grafo logístico — 422 `PLACEMENT_INVALID` si no se cumplen. El coste de
        construcción se asienta en el ledger.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildingCreate'
      responses:
        '201':
          description: Edificio en construcción (`under_construction`).
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Building'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: 'Requisitos de emplazamiento no cumplidos (`PLACEMENT_INVALID`) o fondos insuficientes (`INSUFFICIENT_FUNDS`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
  /world/buildings/{buildingId}:
    parameters:
      - name: buildingId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/BuildingId'
    get:
      operationId: getBuilding
      tags: [world, production]
      summary: Detalle de un edificio
      responses:
        '200':
          description: Edificio.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Building'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
    patch:
      operationId: updateBuilding
      tags: [world, production]
      summary: Configurar un edificio
      description: Cambia la receta activa (con `changeover_seconds` de la receta) o inicia mantenimiento.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildingUpdate'
      responses:
        '200':
          description: Edificio actualizado.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Building'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: 'Receta no soportada por el tipo de edificio, o `min_city_level` no alcanzado por la ciudad cercana.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
  /world/buildings/{buildingId}/upgrade:
    post:
      operationId: upgradeBuilding
      tags: [world, production]
      summary: Mejorar el nivel de un edificio
      description: |
        Progresión **por escala, no por desbloqueo** (GDD 6.3): sube el nivel (líneas
        paralelas, velocidad, eficiencia) con coste no lineal según la `level_curve`
        del tipo. El coste se asienta en el ledger.
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/BuildingId'
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Mejora iniciada.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Building'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Nivel máximo ya alcanzado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Fondos insuficientes (`INSUFFICIENT_FUNDS`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
  /world/buildings/{buildingId}/inventory:
    get:
      operationId: getBuildingInventory
      tags: [world, production]
      summary: Inventario físico de un edificio
      description: |
        Vista **física** del stock por producto. El stock comprometible (libre/reservado)
        se consulta en las cuentas del ledger (`/ledger/accounts`); la reconciliación
        física↔contable es periódica (ADR-004).
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/BuildingId'
      responses:
        '200':
          description: Inventario por producto.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryItem'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        Token de sesión emitido por `POST /auth/sessions`. El servidor solo almacena su
        hash. Humanos y bots usan el mismo esquema (credenciales por cuenta); los bots
        acceden por un camino de red interno multiplexado con la misma superficie de API.
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Clave de idempotencia para **reintentos seguros de comandos que mueven
        valor**: si el cliente reintenta con la misma clave, el servidor reproduce
        la misma respuesta del primer intento — nunca hay doble ejecución.
      schema:
        type: string
        format: uuid
    cursor:
      name: cursor
      in: query
      description: Cursor opaco de paginación devuelto en `meta.next_cursor`.
      schema:
        type: string
    limit:
      name: limit
      in: query
      description: Tamaño máximo de página.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    BadRequest:
      description: Comando malformado o unidades inválidas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Sesión ausente o expirada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Comandar recursos de otra corporación (autorización por propiedad), o vehículo SELLADO durante handoff.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Entidad inexistente (UUID no resuelto).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Conflicto de estado (aceptación sobre publicación agotada, cooldown, stock ya reservado).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    TooManyRequests:
      description: Rate limit excedido — **idéntico para humanos y bots** (decisión de balance, no solo de protección).
      headers:
        Retry-After:
          description: Segundos hasta poder reintentar.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Error inesperado. Nunca deja garantías a medio bloquear — toda operación sobre dinero/stock es atómica en el ledger (todo-o-nada).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Maintenance:
      description: |
        Servicio temporalmente no disponible; SIEMPRE reintentable con la MISMA petición.
        Dos causas, distinguibles por el `code`:

        - `MAINTENANCE_WINDOW`: ventana de mantenimiento diaria (sim-time congelado de forma coordinada).
        - `SERIALIZATION_CONFLICT`: la operación chocó con otras concurrentes sobre las mismas
          cuentas del ledger y agotó su presupuesto de reintentos en el servidor. La transacción se
          revirtió ENTERA —no queda ningún efecto parcial, ni garantías a medio bloquear— así que
          reenviar la misma petición tras `Retry-After` es seguro y es lo esperado del cliente.
      headers:
        Retry-After:
          description: Segundos estimados hasta poder reintentar (fin de la ventana, o drenaje de la ráfaga de escrituras).
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            maintenance:
              value:
                error:
                  code: MAINTENANCE_WINDOW
                  message: El mundo está en su ventana de mantenimiento diaria; el sim-time está congelado
                  details: { retry_after_seconds: 900 }
            serialization_conflict:
              value:
                error:
                  code: SERIALIZATION_CONFLICT
                  message: la operación chocó con otras concurrentes y no se aplicó; reintenta la misma petición
                  details: { retry_after_seconds: 1 }
  schemas:
...
    AccountId:
      type: string
      format: uuid
      description: UUIDv7 de corporación.
      examples: [01981c5e-7d2a-7f3b-9e41-a2c4d6e8f012]
...
    RegionId:
      type: string
      format: uuid
      description: UUIDv7 de región.
    ProductId:
      type: string
      format: uuid
      description: UUIDv7 de producto.
    BuildingTypeId:
      type: string
      format: uuid
      description: UUIDv7 de tipo de edificio.
    RecipeId:
      type: string
      format: uuid
      description: UUIDv7 de receta.
...
    ConcessionId:
      type: string
      format: uuid
      description: UUIDv7 de concesión de suelo.
...
    BuildingId:
      type: string
      format: uuid
      description: UUIDv7 de edificio.
...
    BuildingStatus:
      type: string
      enum: [under_construction, operational, damaged, in_maintenance, abandoned, seized]
...
    Meta:
      type: object
      description: Metadatos comunes a toda respuesta exitosa.
      required: [sim_time, server_time]
      properties:
        sim_time:
          type: string
          pattern: '^[0-9]{1,4}-[0-9]{3}-[0-9]{2}:[0-9]{2}$'
          description: 'Sim-time actual del mundo en formato legible `AÑO-DÍA-HH:MM` (p. ej. año 360 de juego, día 45).'
          examples: ["360-045-12:30"]
        sim_time_seconds:
          $ref: '#/components/schemas/SimTime'
          description: Sim-time actual en segundos desde el génesis (forma canónica para clientes).
        server_time:
          type: string
          format: date-time
          description: Wall-clock del servidor (solo informativo — ninguna regla de juego depende de él).
        next_cursor:
          type: string
          description: Cursor de la página siguiente en respuestas paginadas; ausente si no hay más resultados.
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
      examples:
        - error:
            code: INSUFFICIENT_COLLATERAL
            message: La garantía disponible no cubre la publicación solicitada
            details: { required: "1000", available: "740" }

    # ── Auth ──
...
    Building:
      type: object
      required: [id, owner_account_id, region_id, concession_id, building_type_id, footprint, level, status, condition_pct, fuel_stock]
      properties:
        id:
          $ref: '#/components/schemas/BuildingId'
        owner_account_id:
          $ref: '#/components/schemas/AccountId'
        region_id:
          $ref: '#/components/schemas/RegionId'
        concession_id:
          $ref: '#/components/schemas/ConcessionId'
          description: El edificio pertenece a la corporación; el suelo es siempre concesión del sistema.
        building_type_id:
          $ref: '#/components/schemas/BuildingTypeId'
        footprint:
          $ref: '#/components/schemas/GeoPolygon'
        level:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/BuildingStatus'
        active_recipe_id:
          $ref: '#/components/schemas/RecipeId'
        condition_pct:
          type: integer
          minimum: 0
          maximum: 100
          description: Estado de conservación (degrada con el impago de mantenimiento).
        fuel_stock:
          $ref: '#/components/schemas/StockQty'
          description: Almacén de combustible local — sin combustible, la producción pausa.
        updated_at_sim:
          $ref: '#/components/schemas/SimTime'
        created_at:
          type: string
          format: date-time
    BuildingCreate:
      type: object
      required: [building_type_id, concession_id, footprint]
      properties:
        building_type_id:
          $ref: '#/components/schemas/BuildingTypeId'
        concession_id:
          $ref: '#/components/schemas/ConcessionId'
          description: Concesión propia sobre la que se asienta.
        footprint:
          $ref: '#/components/schemas/GeoPolygon'
          description: Huella dentro de la parcela; validada contra las `placement_rules` del tipo y el grafo logístico.
    BuildingUpdate:
      type: object
      minProperties: 1
      properties:
        active_recipe_id:
          oneOf:
            - $ref: '#/components/schemas/RecipeId'
            - type: 'null'
          description: Receta activa (aplica el `changeover_seconds` de la receta); `null` para detener la línea.
        start_maintenance:
          type: boolean
          description: Inicia mantenimiento programado (`in_maintenance`).
    InventoryItem:
      type: object
      required: [building_id, product_id, quantity, updated_at_sim]
      properties:
        building_id:
          $ref: '#/components/schemas/BuildingId'
        product_id:
          $ref: '#/components/schemas/ProductId'
        quantity:
          $ref: '#/components/schemas/StockQty'
          description: Stock físico total en el edificio (la partición libre/reservado vive en el ledger).
        updated_at_sim:
          $ref: '#/components/schemas/SimTime'
...